Added save/load and new algo
Added ability to save/load grid states and I've included 3 test grids in
the samplegrids folder. I've also added a new algorithm that continue on
in the same direction until obstructed and will backtrack if it runs out
of fresh moves. So far it solves everything more quickly than the
'random' and 'random - prefer unexplored' algos.