New algo and minor UI improvements
Simplified the way new algos are added to the GUI. Added an updating
window title with the file path of the loaded gridstate file.

Added shortcutter algorithm which will cut to adjacent grey blocks
instead of backtracking if they are in our move history between our
current grid and the last grid that bordered a white or blue box. This
mode just takes the first shortcut it can. It's not perfect but it does
better than a non-shortcutter.

Future modes are by shortcut size and by a distance metric. Shortcut
size
will preference the largest shortcut, skipping back the most steps
regardless of whether it gets us closer to our last decision point or
not. The distance metric will choose the move that physically moves us
closer to the last decision point.

Also some new test gridstates.