New algos + bug fix
Fixed a bug in the new shortcutter algorithm that would blow up if the
snake looped on itself (see new testgrid12, it is the smallest example
that could cause the bug to show up). Also added two new algorithms
based on the shortcutter. These two new ones will either try to use the
largest shortcut or take the shortcut that gets closer to the last
decision point. These two prove to be better than the standard
shortcutter although in some instances they actually perform worse.