Bug fixes
Had to fix two annoying bugs. First was the failure to load old grids
due to the change in the way we handle objectives. Fixed this by adding
an "init mode" to setGridItem that bypasses the old mode check (which
isn't needed on first load of a gridstate file). Second bug was the
double counting of objectives.