Pathfinder 1
First rough draft of the pathfinder algorithm. In principle pathfinders process is very simple. It proceeds in the same direction down a path until it comes to a split. It takes one direction and saves the other for later. When it comes upon a dead end, it jumps back to that split and continues on in the next available direction. In theory it should be able to follow all possible paths until no more are available.

This works very well on closed, 1-block wide paths. But if the path is wider than 1 block the theory breaks down and it no longer functions optimally. I've been trying to fix this for a couple days now but decided just to put a pin it it here. For now pathfinder 1 is for closed start-to-finish paths that are only 1 block wide. Also included some more test grids.