Better pathfinder
I realized I shouldnt be keeping decision points made between the last bookmark and the current position in a test branch that becomes larger than the already-best successfull path. It now functions great on testgrid_2 with finish mark in the lower right corner (which was previsouly difficult for it, would get stuck in endless loops in the corners).

Also fixed a bug in resetting the grids and added a way to show where the algorithm is current searching.