Added secondary mode w/ metric
Added a simple distance check with metric enabled to choose the path that leads towards the objective first. I has a tendency to get stuck if the objective and the seeker are separated by a wall with the entrace far away from the objective. In those instances the first choice isn't the best choice and it has to brute force to find the path. Besides those cases the speed is much improved over the previous approach.

Going to try the A* algorithm next, it should be much better than what we have in terms of its execution time.