A* finished
Implemented a basic A* system and damn is it so much faster and better than pathfinder. Haven't been able to trip this one up so far with any fancy gridstates. The pathfinder algos could be tricked in certain ways that vastly increased the time it would take to complete.

The next task I think will be to try and adapt this A* to select the most efficient path between all the objectives (traveling salesman problem). Or maybe I'll just find an algorithm thats specifically for that and not try and do it myself, we'll see.

Also fixed a bug that prevented the found paths from being played out in realtime when the stop buttom was pressed mid-search.